Feat: Implement replying to comments
Each comment gets a "Reply" button if you're logged in. Clicking it POSTs to
?/reply_to_comment, which checks the parentId from the <input type="hidden">
field exists, then returns the comment info.

That is then set as form.replyTo, and the UI updates to change the "Add a
comment:" label to "Reply to @username".

Progressive enhancement is implemented through use:enhance, so the page doesn't
reload if JavaScript is enabled. A callback is provided, which calls an
onReply(ClientSideComment) handler. We pass this handler down from the page;
it focuses the comment entry field once its label and hidden parentId field
have been changed.