feat: Display threaded comments
Comments are retrieved from the DB as a flat list, and are then turned into an
array of trees by Comment.constructThreads().

The client recursively alternates between <CommentList> and <Comment> elements
to render the tree; each comment is rendered within a <li> element, inside of
an <ul>.

No visual distinction between levels has been added yet, but the tree is
correctly communicated semantically to screen readers.